chore: reorganize routes engine rules
