+++ src/proxy/routes-engine/index.ts
@@ -1,0 +1,7 @@
+// The routes engine implemetation is (and should always be) aligned with the documentation
+// see: https://docs.microsoft.com/en-us/azure/static-web-apps/configuration
+
+export { customRoutes } from "./rules/customRoutes";
+export { globalHeaders } from "./rules/globalHeaders";
+export { mimeTypes } from "./rules/mimeTypes";
+export { responseOverrides } from "./rules/responseOverrides";
