+++ src/proxy/routes-engine/rules/globalHeaders.spec.ts
@@ -1,6 +1,6 @@
-import { processGlobalHeaders } from "./routes-engine";
+import { globalHeaders } from "./globalHeaders";
 
-describe("processGlobalHeaders()", () => {
+describe("globalHeaders()", () => {
   let req: any;
   let res: any;
   let userConfig: SWAConfigFileGlobalHeaders;
@@ -16,7 +16,7 @@ describe("processGlobalHeaders()", () => {
   });
 
   it("should not set headers if empty config", async () => {
-    await processGlobalHeaders(req, res, userConfig);
+    await globalHeaders(req, res, userConfig);
 
     expect(res.setHeader).not.toHaveBeenCalled();
   });
@@ -25,7 +25,7 @@ describe("processGlobalHeaders()", () => {
     userConfig = {
       "Cache-Control": "public, max-age=604800",
     };
-    await processGlobalHeaders(req, res, userConfig);
+    await globalHeaders(req, res, userConfig);
 
     expect(res.setHeader).toHaveBeenCalledWith("Cache-Control", "public, max-age=604800");
   });
@@ -34,7 +34,7 @@ describe("processGlobalHeaders()", () => {
     userConfig = {
       "Cache-Control": "",
     };
-    await processGlobalHeaders(req, res, userConfig);
+    await globalHeaders(req, res, userConfig);
 
     expect(res.removeHeader).toHaveBeenCalledWith("Cache-Control");
     expect(res.setHeader).not.toHaveBeenCalled();
